# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API MathEasy",
        "description": "API per gestionar operacions i recursos de MathEasy amb Laravel i autenticació Bearer",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://192.168.1.39:8000/api",
            "description": "Servidor local de desenvolupament"
        }
    ],
    "paths": {
        "/alumnes": {
            "get": {
                "tags": [
                    "Alumnes"
                ],
                "summary": "Llistar tots els alumnes",
                "description": "Retorna una llista amb tots els alumnes registrats",
                "operationId": "136a14845978f96f58e89b302a2add0e",
                "responses": {
                    "200": {
                        "description": "Llista d'alumnes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Alumne"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Usuari no autenticat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Alumnes"
                ],
                "summary": "Crear un nou alumne",
                "description": "Crea un alumne nou amb les dades proporcionades",
                "operationId": "f402569706d709e193d14dd37fabd227",
                "requestBody": {
                    "description": "Dades de l'alumne",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AlumneInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Alumne creat correctament",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Alumne"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dades no vàlides"
                    },
                    "401": {
                        "description": "Usuari no autenticat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/alumnes/{id}": {
            "get": {
                "tags": [
                    "Alumnes"
                ],
                "summary": "Mostrar un alumne",
                "description": "Retorna les dades d'un alumne específic pel seu ID",
                "operationId": "9080ee90af4f7973038fd9b97d773847",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'alumne",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alumne trobat correctament",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Alumne"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Alumne no trobat"
                    },
                    "401": {
                        "description": "Usuari no autenticat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Alumnes"
                ],
                "summary": "Actualitzar dades d'un alumne",
                "description": "Actualitza les dades personals d'un alumne autenticat.",
                "operationId": "d090b6e280b0588cc1e834e7f425edb9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'alumne",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dades a actualitzar de l'alumne",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AlumneUpdateInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Alumne actualitzat correctament",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Alumne"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticat"
                    },
                    "404": {
                        "description": "Alumne no trobat"
                    },
                    "422": {
                        "description": "Dades no vàlides"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/loggin": {
            "post": {
                "tags": [
                    "Alumnes"
                ],
                "summary": "Iniciar sessió d'un alumne",
                "description": "Comprova el nom d'usuari i la contrasenya. Retorna les dades de l'alumne si és correcte.",
                "operationId": "eed7e793b3ea10acda6b6bd66e2a2c56",
                "requestBody": {
                    "description": "Dades d'inici de sessió",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuari autenticat correctament",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Usuari i/o contrasenya incorrectes"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tokenLoggin": {
            "post": {
                "tags": [
                    "Alumnes"
                ],
                "summary": "Validar token",
                "operationId": "df7f3482c82eba4de12be93d284b8baa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TokenLoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token vàlid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Alumne"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token invàlid"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/perfilImage": {
            "post": {
                "tags": [
                    "Alumnes"
                ],
                "summary": "Pujar imatge de perfil",
                "operationId": "4965a80c46031b213507fe005960509c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProfileImageUpload"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Imatge pujada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "path": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/perfilImageEdit": {
            "post": {
                "tags": [
                    "Alumnes"
                ],
                "summary": "Editar imatge de perfil",
                "operationId": "b331dbb0973bde12ef495a85d840b8bb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProfileImageEdit"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Imatge actualitzada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "path": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/alumnes/experiencia/{id}": {
            "get": {
                "tags": [
                    "Alumnes"
                ],
                "summary": "Obtenir experiència",
                "description": "Retorna dades d'experiència d'un alumne",
                "operationId": "b012d53cd33bdbf8515f7f26b41c9d23",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Experiència trobada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Experiencia"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No trobada"
                    },
                    "401": {
                        "description": "Usuari no autenticat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Alumnes"
                ],
                "summary": "Actualitzar experiència",
                "description": "Actualitza nivell, XP i medalles",
                "operationId": "4945977539b9bec25501429e3271be05",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'experiencia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExperienciaUpdateInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualitzat correctament",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Experiencia"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticat"
                    },
                    "404": {
                        "description": "Alumne no trobat"
                    },
                    "422": {
                        "description": "Dades no vàlides"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "LoginResponse": {
                "title": "Login Response",
                "properties": {
                    "Resposta": {
                        "properties": {
                            "Alumne": {
                                "$ref": "#/components/schemas/Alumne"
                            },
                            "token": {
                                "type": "string",
                                "example": "1|abc123token"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "Alumne": {
                "title": "Alumne",
                "description": "Representa un alumne, heretat de Usuaris",
                "properties": {
                    "id": {
                        "description": "ID de l'alumne",
                        "type": "integer"
                    },
                    "Nom": {
                        "description": "Nom de l'usuari",
                        "type": "string"
                    },
                    "Cognoms": {
                        "description": "Cognoms de l'usuari",
                        "type": "string"
                    },
                    "Password": {
                        "description": "Password de l'usuari",
                        "type": "string",
                        "format": "password"
                    },
                    "Nom_Usuari": {
                        "description": "Nom d'usuari dins del curs",
                        "type": "string"
                    },
                    "Curs": {
                        "description": "Curs de l'alumne",
                        "type": "string"
                    },
                    "Experiencia": {
                        "description": "Experiència de l'alumne",
                        "type": "integer",
                        "format": "int32"
                    },
                    "ProfilePicturePath": {
                        "description": "Imatge de perfil",
                        "type": "string",
                        "example": "images/avatar.jpg"
                    }
                },
                "type": "object"
            },
            "AlumneInput": {
                "title": "Entrada Alumne",
                "description": "Dades necessàries per crear un alumne",
                "properties": {
                    "Nom": {
                        "description": "Nom de l'usuari",
                        "type": "string"
                    },
                    "Cognoms": {
                        "description": "Cognoms de l'usuari",
                        "type": "string"
                    },
                    "Password": {
                        "description": "Password de l'usuari",
                        "type": "string",
                        "format": "password"
                    },
                    "Nom_Usuari": {
                        "description": "Nom d'usuari dins del curs",
                        "type": "string"
                    },
                    "Curs": {
                        "description": "Curs de l'alumne",
                        "type": "string"
                    },
                    "Experiencia": {
                        "description": "Experiència de l'alumne",
                        "type": "integer",
                        "format": "int32"
                    },
                    "ProfilePicturePath": {
                        "description": "Imatge de perfil",
                        "type": "string",
                        "example": "images/avatar.jpg"
                    }
                },
                "type": "object"
            },
...
            "LoginRequest": {
                "title": "LoginRequest",
                "description": "Dades necessàries per iniciar sessió",
                "properties": {
                    "Nom_Usuari": {
                        "description": "Nom d'usuari de l'alumne",
                        "type": "string"
                    },
                    "Password": {
                        "description": "Contrasenya de l'alumne",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Introduïu el token Bearer JWT per autenticar-vos",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}